15687 - move routes into v1 folder
