+++ pay-api/src/pay_api/resources/v1/__init__.py
@@ -16,24 +16,24 @@ from typing import Optional
 
 from flask import Flask
 
-from ..account import bp as account_bp
-from ..account_statements import bp as account_statements_bp
-from ..account_statements_notifications import bp as account_notifications_bp
-from ..account_statements_settings import bp as account_settings_bp
-from ..bank_accounts import bp as bank_accounts_bp
-from ..code import bp as code_bp
-from ..distributions import bp as distributions_bp
-from ..fas import fas_refund_bp, fas_routing_slip_bp
-from ..fee import bp as fee_bp
-from ..fee_schedule import bp as fee_schedule_bp
-from ..invoice import bp as invoice_bp
-from ..invoice_receipt import bp as invoice_receipt_bp
-from ..invoices import bp as invoices_bp
-from ..meta import bp as meta_bp
+from .account import bp as account_bp
+from .account_statements import bp as account_statements_bp
+from .account_statements_notifications import bp as account_notifications_bp
+from .account_statements_settings import bp as account_settings_bp
+from .bank_accounts import bp as bank_accounts_bp
+from .code import bp as code_bp
+from .distributions import bp as distributions_bp
+from .fas import fas_refund_bp, fas_routing_slip_bp
+from .fee import bp as fee_bp
+from .fee_schedule import bp as fee_schedule_bp
+from .invoice import bp as invoice_bp
+from .invoice_receipt import bp as invoice_receipt_bp
+from .invoices import bp as invoices_bp
+from .meta import bp as meta_bp
 from ..ops import bp as ops_bp
-from ..payment import bp as payment_bp
-from ..refund import bp as refund_bp
-from ..transaction import bp as transaction_bp
+from .payment import bp as payment_bp
+from .refund import bp as refund_bp
+from .transaction import bp as transaction_bp
 
 
 class V1Endpoint:  # pylint: disable=too-few-public-methods,
